Update list of created folders and files
The list of files and folders created is updated, adding the missing "var" folder;
text is used instead of an image - it is more consistent with the other page
(Install without composer).

Update formatting of code block. Also added the .gitignore file to the list.

Remove no longer required image.