Use codeblock instead of inline code
Use a codeblock instead of inline code. Previously,
the line was truncated.

Now, we use a real code-block. This also makes the command more
readable and it is easier to see what is the difference to the
next command which is very similar.

Since there are several code blocks on that page for shell we
use shell hightlight by default.

We use correct lexer for Apache and Ngxinx configuration.