Use consistent naming for public directory (#179)
By default, "public" is used in Composer installations. The page
about installing TYPO3 without Composer uses "htdocs". By
changing this to "public", we already have a directory structure
that is more consistent with the one used in Composer installations.
This might make it a little more easier to migrate later.

Also, this is more consistent with the rest of the guide which uses
"public".

Resolves: #178