Clarify execution of upgrade wizards
Previous text was slightly misleading: It proposed to not execute the optional
wizard. However, from the GUI and the text is not clear that you should
actually click on the buttton "Execute" and then choose "No, do not execute".

Otherwise the upgrade wizard will be displayed in the list of to be executed
wizards and the upgrade will not be considered as done.

This change proposed to execute the wizard with the "Do not execute" option.