Update Index.rst
Comparing https://semver.org/ with https://docs.typo3.org/m/typo3/guide-installation/main/en-us/Minor/Index.html#minor shows that the word "minor" is not correct. It has to be clarified as a patch level update. If I see this correctly the update for a minor change is not described in the documentation since the wrong minor article is the bugfix upgrade article. Please correct me if I missed something.