Documentation Inconsistency: The level of described change is patch not minor. Plz Update Index.rst (#322)
Comparing https://semver.org/ with https://docs.typo3.org/m/typo3/guide-installation/main/en-us/Minor/Index.html#minor shows that the word "minor" is not correct. It has to be clarified as a patch level update. If I see this correctly the update for a minor change is not described in the documentation since the wrong minor article is the bugfix upgrade article. Please correct me if I missed something.

---------

Co-authored-by: Chris Müller <2566282+brotkrueml@users.noreply.github.com>